Finish Lab 40

+++ Lab40/Moveable.java
@@ -1,11 +1,11 @@
 public interface Moveable
 {
     public void setPos( int x, int y);
-    public void setX( int x );
-    public void setY( int y );
+    public void setX(double x );
+    public void setY(double y );
 
-    public int getX();
-    public int getY();
+    public double getX();
+    public double getY();
 
 	 public int getWidth();
 	 public int getHeight();
